Move todos-related code inside features/todos directory
