+++ src/features/todos/components/TodoListEmpty.svelte
@@ -2,7 +2,7 @@
 import { createEventDispatcher } from 'svelte';
 import { dndzone, TRIGGERS } from 'svelte-dnd-action';
 import Button from '@components/Button.svelte';
-import TodoItem from '@components/TodoItem.svelte';
+import TodoItem from './TodoItem.svelte';
 
 export let text;
